Version 8.01 Touchpad macros now work without also clicking) Mouse Buttons are now toggleable
Version 8
More mapping options!
Toggle option: Press a button once and the key will be held down until the button is pressed again
Macro option: Assign multiple keys or mouse buttons to a single button! Just check Macro and select the buttons in the order you want them to be pressed (maximum 5)
Light bar even closer to default bluetooth when stopping
When disconnecting Controller via PS button and Options, if PS is assigned to a key or keys, the key is let go.
If Idle timeout is on, the light bar idles fades when fully charged (but doesn't disconnect)
DS4 Icon always shows, notifications only at startup, and when the main window is unfocused.
In options the tap, scroll, and slide are unchecked at start if any of them were turned off before
For always rainbow mode, the cycle goes in reverse when charging, remains dim if checked.

Version 7.8
Mouse Sensitivity now works with analog sticks as well as buttons and triggers
Option for using the controller input to change it's mapping in options
Fix for lightbar not remembering the color
Disable edit/delete buttons until controllers finish connecting