Version 1.4.2
Special Actions: press a control or a set of controls to perform a
macro, launch a program/file, or load a different profile. Create new
actions in profile options, and check which ones you want the selected
profile to use.
Set an analog curve for the left and right stick, see the changes live
in the controller readout
Option to set the close button to minimize DS4Windows, and truly close
it by the notification icon
Added Spanish Translations (added credits to the translations sheets so
users can be properly thanked if wanted to be)
Sticks Deadzone reworked from being a square to radial as the controller
readout actually shows
Fix pressing keyboard not setting an action for controls
Fix Tilts and Swipes controls not showing text in the control list
Fix program crashing when uses x360 macros with shift controls
Fix for trying to edit a profile from the context menu after never
restoring DS4Windows form start.
Fix for settings low flash color
Now running with startup will always start minimized, even if start
minimized is unchecked