Add AutoProfile target EXE to Hidhide
Changes CheckHidHidePresence() slightly to be able to add specified EXEs to HidHide's whitelist.

This is backwards compatible. Default values are used (probably poorly) to detect D4Ws initialization, and not interfere with adding itself to HidHide on startup. A RemoveExe Bool is used to check if the program should Add or Remove the chosen EXE.